python-flask-api/requirements.txt
Flask==1.0.1
Jinja2==2.10
Keras==2.0.5
MarkupSafe==1.0
Pillow==5.0.0
PyYAML==3.12
Theano==1.0.1
Werkzeug==0.14.1
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
cycler==0.10.0
dill==0.2.7.1
docutils==0.14
future==0.16.0
grpcio==1.12.1
gunicorn==19.8.1
h5py==2.7.0
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.2
oauth2client==3.0.0
opencv-python==3.2.0.8
pandas==0.19.1
ply==3.8
protobuf==3.5.2.post1
psutil==5.4.6
pyasn1==0.4.3
pyasn1-modules==0.2.1
pyparsing==2.2.0
pyreadline==2.1
python-dateutil==2.7.1
pytz==2018.3
requests==2.19.1
rsa==3.4.2
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
statistics==1.0.3.5
tensorflow==1.1.0
urllib3==1.23
jsonpickle==0.9.6
